requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
elastic-transport==8.4.0
elasticsearch==8.8.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
orjson==3.9.0
pandas==2.0.2
pydantic==1.10.8
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
